pyproject.toml
[tool.poetry]
name = "seagoat"
version = "0.1.2"
description = "A semantic-code search engine"
authors = ["Daniel Kantor <git@daniel-kantor.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.scripts]
gt = "seagoat.main:seagoat"
seagoat = "seagoat.main:seagoat"
seagoat-server = "seagoat.server:cli"

[tool.poetry.dependencies]
python = "^3.11"
chromadb = "^0.3.26"
gitpython = "^3.1.31"
tqdm = "^4.65.0"
appdirs = "^1.4.4"
click = "^8.1.3"
prompt-toolkit = "^3.0.38"
blessed = "^1.20.0"
pygments = "^2.15.1"
ripgrepy = "^2.0.0"
nest-asyncio = "^1.5.6"
flask = "^2.3.2"
requests = "^2.31.0"

[virtualenvs]
in-project = true
create = true

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
pylint = "^2.17.4"
black = "^23.3.0"
pre-commit = "^3.3.3"
pyright = "^1.1.314"
pytest-watch = "^4.2.0"
freezegun = "^1.2.2"
syrupy = "^4.0.4"
pytest-ordering = "^0.6"
pytest-asyncio = "^0.21.0"
ipython = "^8.14.0"
pytest-flask = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venvPath = "."
venv = ".venv"
include = ["seagoat"]
exclude = ["**/node_modules",
    "**/__pycache__",
]
ignore = ["typings"]

reportMissingImports = false
reportMissingTypeStubs = false

pythonVersion = "3.11"
pythonPlatform = "Linux"
strict = []

executionEnvironments = [
  { root = "src" }
]
